refactoring

+++ app/Http/Controllers/Api/HomeImageController.php
@@ -87,16 +87,27 @@ class HomeImageController extends Controller
   /**
    * Remove the specified resource from storage.
    *
-   * @param  HomeImage $image
+   * @param  string $image
    * @return \Illuminate\Http\Response
    */
-  public function destroy(HomeImage $image)
+  
+  public function destroy($image)
   {
-    $image->delete();
+    // Delete image from database
+    $image = $this->image->where('name', '=', $image)->first();
+    
+    if ($image)
+    {
+      $image->delete();
+    }
+
+    // Delete file from storage
     $directories = Storage::allDirectories('public');
     foreach($directories as $d)
     {
       Storage::delete($d . '/'. $image->name);
     }
+    
+    return response()->json('successfully deleted');
   }
 }
